requirements.txt
# Add your project dependencies here
Flask==2.3.3
requests==2.31.0
hashlib
hmac
base64
json
urllib.parse
qrcode[pil]==7.4.2
Pillow==10.0.1
cryptography==41.0.7
